update to Chromium 125.0.6422.26

+++ content/browser/web_package/signed_exchange_cert_fetcher.cc
@@ -9,7 +9,6 @@
 #include "base/functional/bind.h"
 #include "base/metrics/histogram_macros.h"
 #include "base/numerics/safe_conversions.h"
-#include "base/strings/string_piece.h"
 #include "base/strings/stringprintf.h"
 #include "base/task/sequenced_task_runner.h"
 #include "base/task/single_thread_task_runner.h"
@@ -186,7 +185,7 @@ void SignedExchangeCertFetcher::OnHandleReady(MojoResult result) {
   TRACE_EVENT0(TRACE_DISABLED_BY_DEFAULT("loading"),
                "SignedExchangeCertFetcher::OnHandleReady");
   const void* buffer = nullptr;
-  uint32_t num_bytes = 0;
+  size_t num_bytes = 0;
   MojoResult rv =
       body_->BeginReadData(&buffer, &num_bytes, MOJO_READ_DATA_FLAG_NONE);
   if (rv == MOJO_RESULT_OK) {
